Final fixes and documentation
* Changed vote count type to long (because the total vote count can go
up to 2^32 - 1);
* Added documentation and improved comments in code;
* Removed temporary test input files from solution.

+++ Solutions/musashibg/src/Party.cs
@@ -4,6 +4,10 @@ using System.Text;
 
 namespace MandateCalculator
 {
+	/// <summary>
+	/// Съхранява информацията за една партия/коалиция, заредени от
+	/// Parties.txt.
+	/// </summary>
 	public class Party
 	{
 		public Party()
@@ -13,18 +17,49 @@ namespace MandateCalculator
 			MandateAssignments = new List<MandateAssignment>();
 		}
 
+		/// <summary>
+		/// Номер на партията/коалицията.
+		/// </summary>
 		public int PartyId { get; set; }
 
+		/// <summary>
+		/// Име на партията/коалицията.
+		/// </summary>
 		public string Name { get; set; }
 
+		/// <summary>
+		/// Колекция от всички кандидати, регистрирани в кандидатската листа на
+		/// партията/коалицията.
+		/// </summary>
 		public List<Candidate> Candidates { get; private set; }
 
+		/// <summary>
+		/// Колекция от всички групи от действителни гласове, подадени във
+		/// всеки многомандатен изборен район за партията/коалицията. Ключът на
+		/// колекцията е номерът на многомандатния изборен район, в който са
+		/// подадени гласовете.
+		/// </summary>
 		public Dictionary<int, VoteBatch> VoteBatches { get; private set; }
 
+		/// <summary>
+		/// Брой мандати, които партията/коалицията трябва да получи на
+		/// национално ниво (по чл. 16).
+		/// </summary>
 		public int NationalMandateCount { get; set; }
 
+		/// <summary>
+		/// Колекция от обекти с информация за разпределените на
+		/// партията/коалицията мандати във всеки многомандатен изборен район
+		/// (по чл. 21 - 27).
+		/// </summary>
 		public List<MandateAssignment> MandateAssignments { get; private set; }
 
+		/// <summary>
+		/// Връща символен низ с информацията за партията/коалицията във вид,
+		/// удобен за отпечатване.
+		/// </summary>
+		/// <returns>Символен низ с информацията за
+		/// партията/коалицията.</returns>
 		public override string ToString()
 		{
 			var builder = new StringBuilder();
@@ -34,11 +69,22 @@ namespace MandateCalculator
 			return builder.ToString();
 		}
 
-		public int GetTotalVoteCount()
+		/// <summary>
+		/// Изчислява общия брой действителни гласове, подадени за
+		/// партията/коалицията във всички многомандатни изборни райони.
+		/// </summary>
+		/// <returns>Общ брой подадени действителни гласове.</returns>
+		public long GetTotalVoteCount()
 		{
 			return VoteBatches.Values.Sum(vb => vb.VoteCount);
 		}
 
+		/// <summary>
+		/// Изчислява общия брой мандати, текущо разпределени на
+		/// партията/коалицията в многомандатните изборни райони.
+		/// </summary>
+		/// <returns>Общ брой мандати, разпределени на регионално
+		/// ниво.</returns>
 		public int GetAssignedMandateCount()
 		{
 			int mandateCount = 0;
